Improve styling and organize CSS
