+++ public/index.html
@@ -4,6 +4,13 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Kioskito Admin Dashboard</title>
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css"
+      integrity="sha512-dO1eWnXniCx9aRUZ3dzpQttlhE2wAPOQgpdnXl16rYPD+s/COM24kTx5c5NyEJD7BqXc9E+u6KDAdM8wEJXKVA=="
+      crossorigin="anonymous"
+      referrerpolicy="no-referrer"
+    />
   </head>
   <body>
     <div id="root"></div>
