--- src/components/Configuration.css
@@ -1,6 +1,0 @@
-.configuration {
-  background-color: #fff;
-  border-top: 1px solid #ccc;
-  padding: 20px;
-  margin-top: auto;
-}
