--- src/components/Login.css
@@ -1,32 +1,0 @@
-.login-container {
-  max-width: 300px;
-  margin: 50px auto;
-  padding: 20px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  background: #fff;
-}
-
-.login-container form div {
-  margin-bottom: 10px;
-}
-
-.login-container label {
-  display: block;
-  margin-bottom: 5px;
-}
-
-.login-container input {
-  width: 100%;
-  padding: 8px;
-  box-sizing: border-box;
-}
-
-.login-container .error {
-  color: red;
-  margin-bottom: 10px;
-}
-
-.login-container button {
-  padding: 8px 12px;
-}
