+++ src/components/Sidebar.js
@@ -1,14 +1,14 @@
 import React from 'react';
-import './Sidebar.css';
+import '../styles/Sidebar.css';
 
 function Sidebar() {
   return (
     <aside className="sidebar">
       <nav>
         <ul>
-          <li>Inventario</li>
-          <li>Ventas</li>
-          <li>Concesiones</li>
+          <li><i className="fas fa-boxes"></i> Inventario</li>
+          <li><i className="fas fa-cash-register"></i> Ventas</li>
+          <li><i className="fas fa-store"></i> Concesiones</li>
         </ul>
       </nav>
     </aside>
