+++ src/config.js
@@ -1,0 +1,1 @@
+export const API_URL = 'https://kioskito-api.onrender.com';
