+++ src/styles/App.css
@@ -1,6 +1,9 @@
+@import './variables.css';
+
 .app-container {
   display: flex;
   height: 100vh;
+  background-color: var(--bg-color);
 }
 
 .content {
