+++ src/styles/Configuration.css
@@ -1,0 +1,8 @@
+@import './variables.css';
+
+.configuration {
+  background-color: var(--text-color);
+  border-top: 1px solid var(--primary-color);
+  padding: 20px;
+  margin-top: auto;
+}
