+++ src/styles/Login.css
@@ -1,0 +1,46 @@
+@import './variables.css';
+
+.login-container {
+  max-width: 320px;
+  margin: 60px auto;
+  padding: 20px;
+  border-radius: 8px;
+  background: var(--text-color);
+  box-shadow: 0 2px 6px rgba(0,0,0,0.1);
+}
+
+.login-container form div {
+  margin-bottom: 12px;
+}
+
+.login-container label {
+  display: block;
+  margin-bottom: 5px;
+  color: var(--primary-color);
+}
+
+.login-container input {
+  width: 100%;
+  padding: 8px;
+  box-sizing: border-box;
+  border: 1px solid var(--secondary-color);
+  border-radius: 4px;
+}
+
+.login-container .error {
+  color: red;
+  margin-bottom: 10px;
+}
+
+.login-container button {
+  background-color: var(--secondary-color);
+  color: var(--text-color);
+  border: none;
+  padding: 8px 12px;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.login-container button:hover {
+  background-color: var(--accent-color);
+}
