+++ src/styles/Sidebar.css
@@ -1,7 +1,9 @@
+@import './variables.css';
+
 .sidebar {
   width: 220px;
-  background-color: #2f3e46;
-  color: #fff;
+  background-color: var(--primary-color);
+  color: var(--text-color);
   display: flex;
   flex-direction: column;
   padding-top: 20px;
@@ -19,5 +21,9 @@
 }
 
 .sidebar li:hover {
-  background-color: #354f52;
+  background-color: var(--secondary-color);
+}
+
+.sidebar li i {
+  margin-right: 8px;
 }
