+++ src/styles/index.css
@@ -1,0 +1,8 @@
+@import './variables.css';
+
+body {
+  margin: 0;
+  font-family: Arial, Helvetica, sans-serif;
+  background-color: var(--bg-color);
+  color: var(--primary-color);
+}
