+++ src/styles/variables.css
@@ -1,0 +1,7 @@
+:root {
+  --primary-color: #37474f;
+  --secondary-color: #00838f;
+  --accent-color: #4caf50;
+  --bg-color: #eceff1;
+  --text-color: #ffffff;
+}
